requirements.txt
fastapi==0.115.8
itsdangerous==2.2.0
Jinja2==3.1.5
pydantic==2.10.6
python-dotenv==1.0.1
python-multipart==0.0.20
uvicorn==0.34.0
